Add Test summary and push output file to github
The action pytest-results-action will create a nice summary of test time and number of failed/skipped/passed tests.

Added a step to also publish the xml test file to github as an artefact. It will stay for 90 days.
If we want to keep it longer we need our own database. However that can be added later.