Add experimental rig manager firmware and script
Add firmware for an experimental arduino-powered rig manager hardware
that exposes functionalities via an SCPI command set over uart.

The manager controls relays for switching on and off the Crazyflies
and open-drain output to control the buttons

This allows to put Crazyflie in bootloader modes if they get stuck.

Also add a very crude test script to demonstrate the functionality.