Remove the ai deck from testing.
Not the best solution but we dont know what is causing the issues. So removing the ai deck
might give some clues. The failures started happening once we introduced the ai deck delay.
So I want to test the stability of the system when it is gone. Then we can add it back slowly